requirements.txt
spidev>=3.4.0
RPi.GPIO>=0.7.0
pillow>=8.0.0
gpiozero>=1.5.0
lgpio